Fix linting, pull out constants
