+++ lib/byGroup.js
@@ -1,3 +1,5 @@
+'use strict';
+
 module.exports = (data, orderedGroups) => {
   const dataByGroup = {};
   data.forEach(item => {
