+++ lib/getNunjucksEnv.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const nunjucks = require('nunjucks');
 
 module.exports = (name, env, warned) => {
