+++ lib/prepareContext.js
@@ -1,3 +1,18 @@
+'use strict';
+
+/**
+ * SassDoc extras (providing Markdown and other filters, and different way to
+ * index SassDoc data).
+ *
+ * See <https://github.com/SassDoc/sassdoc-extras>.
+ */
+const extras = require('sassdoc-extras');
+const extend = require('extend');
+const fs = require('fs');
+
+const parse = require('./parse.js');
+const byGroup = require('./byGroup');
+
 module.exports = ctx => {
   const def = {
     display: {
