+++ lib/renderHerman.js
@@ -1,3 +1,16 @@
+'use strict';
+
+const Promise = require('bluebird');
+const tinycolor = require('tinycolor2');
+const path = require('path');
+const extend = require('extend');
+
+const render = require('./render.js');
+const copy = require('./assets.js');
+const parse = require('./parse.js');
+
+const { nunjucksEnv, base } = require('../constants');
+
 module.exports = (dest, ctx) => {
   const indexTemplate = path.join(base, 'index.j2');
   const indexDest = path.join(dest, 'index.html');
